requirements.txt
google-genai>=1.0.0
chromadb>=0.4.0
plivo>=4.0.0
composio-gemini>=0.1.0
requests>=2.31.0
flask>=3.0.0
python-dotenv>=1.0.0
rich>=13.0.0
pydantic>=2.0.0
